# docs/swagger.yaml
definitions:
...
  models.Account:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      entries:
        description: Relationships
        items:
          $ref: '#/definitions/models.Entry'
        type: array
      id:
        type: string
      owner:
        type: string
      transfers_from:
        items:
          $ref: '#/definitions/models.Transfer'
        type: array
      transfers_to:
        items:
          $ref: '#/definitions/models.Transfer'
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.Entry:
    properties:
      account:
        $ref: '#/definitions/models.Account'
      accountID:
        type: string
      amount:
        type: integer
      createdAt:
        type: string
      id:
        type: string
    type: object
  models.Transfer:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      from_account:
        $ref: '#/definitions/models.Account'
      from_account_id:
        type: string
      id:
        type: string
      to_account:
        $ref: '#/definitions/models.Account'
      to_account_id:
        type: string
    type: object
  transfer.CreateTransferRequest:
    properties:
      amount:
        type: integer
      from_account_id:
        type: string
      to_account_id:
        type: string
    required:
    - amount
    - from_account_id
    - to_account_id
    type: object
  transfer.CreateTransferResponse:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      from_account_id:
        type: string
      id:
        type: string
      to_account_id:
        type: string
    type: object
  transfer.model:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      from_account:
        $ref: '#/definitions/models.Account'
      from_account_id:
        type: string
      id:
        type: string
      to_account:
        $ref: '#/definitions/models.Account'
      to_account_id:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/v1/transfer:
    get:
      description: Retrieves all transfers for a specific account, filtering by the
        authenticated user's account
      parameters:
      - description: '{''and'': [ {''title'': { ''cont'':''cul'' } } ]}'
        in: query
        name: s
        type: string
      - description: 'fields to select eg: name,age'
        in: query
        name: fields
        type: string
      - description: page of pagination
        in: query
        name: page
        type: integer
      - description: limit of pagination
        in: query
        name: limit
        type: integer
      - description: 'join relations eg: category, parent'
        in: query
        name: join
        type: string
      - collectionFormat: csv
        description: 'filters eg: name||eq||ad price||gte||200'
        in: query
        items:
          type: string
        name: filter
        type: array
      - collectionFormat: csv
        description: 'filters eg: created_at,desc title,asc'
        in: query
        items:
          type: string
        name: sort
        type: array
      - description: ID of the account to filter transfers by
        in: query
        name: account_id
        required: true
        type: string
      - description: 'Direction of transfer: incoming, outgoing, or all (default is
          all)'
        in: query
        name: direction
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/transfer.model'
            type: array
      security:
      - JWT: []
      summary: Get all transfers for an account
      tags:
      - transfer
  /api/v1/transfer/{id}:
    get:
      description: Retrieves a single transfer by its UUID
      parameters:
      - description: uuid of item
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transfer.model'
      security:
      - JWT: []
      summary: Get a transfer by ID
      tags:
      - transfer
  /api/v1/transfer/execute:
    post:
      consumes:
      - application/json
      description: Transfers money from one account to another using a transaction
      parameters:
      - description: Transfer payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/transfer.CreateTransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/transfer.CreateTransferResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Execute a money transfer between accounts
      tags:
      - transfer
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"